+++ common/db/src/main/java/is/codion/common/db/exception/ReferentialIntegrityException.java
@@ -3,19 +3,34 @@
  */
 package is.codion.common.db.exception;
 
+import is.codion.common.db.database.Database.Operation;
+
 import java.sql.SQLException;
+
+import static java.util.Objects.requireNonNull;
 
 /**
  * An exception indication a referential integrity failure
  */
 public final class ReferentialIntegrityException extends DatabaseException {
 
+  private final Operation operation;
+
   /**
    * Instantiates a new {@link ReferentialIntegrityException}
    * @param cause the underlying cause
    * @param message the error message
+   * @param operation the operation causing this exception
    */
-  public ReferentialIntegrityException(SQLException cause, String message) {
+  public ReferentialIntegrityException(SQLException cause, String message, Operation operation) {
     super(cause, message);
+    this.operation = requireNonNull(operation);
+  }
+
+  /**
+   * @return the {@link Operation} causing this exception
+   */
+  public Operation operation() {
+    return operation;
   }
 }